# fix: Compatibility with QuickTime Player on macOS