Implemented the following fix: https://superuser.com/a/820137/813480

Before this was shown and not working by that:

![image](https://user-images.githubusercontent.com/17984549/83201649-41168900-a146-11ea-80e4-f69ad053e169.png)

Reopen of #27 to fix tests too.