@celeryclub 2020-05-29
QuickTime scrubbing is so handy! 🙌 Do you know if this has any performance impact?