@mxschmitt 2020-05-31
@jperl I fixed the tests by changing the minimal PNG image to an actual real image of viewport width and height to 1 pixel.

@celeryclub I did some basic performance testing on my 2019 MacBook Pro, but not sure if we can interpret anything out of these values:

after the change (running three times the script):
- 3.894 seconds
- 4.340 seconds
- 4.673 seconds

file size 2.0 MB

before the change (running three times the script):
- 5.067 seconds
- 4.626 seconds
- 5.028 seconds

file size: 3.0 MB

used this script:

```js
const { chromium } = require('playwright');
const { saveVideo } = require('.');

(async () => {
  const browser = await chromium.launch();
  const context = await browser.newContext();
  const page = await context.newPage();

  const capture = await saveVideo(page, 'video2.mp4');
  await page.goto('https://example.org');
  await page.goto('https://amazon.com');
  await page.goto('https://apple.com');
  await page.goto('https://facebook.com');
  await page.goto('https://github.com');
  await page.goto('https://wikipedia.org');
  const start = new Date()
  await capture.stop()
  console.log(((new Date().getTime() - start.getTime()) / 1000).toFixed(2))
  await browser.close();
})();
```